requirements.txt
matplotlib==3.10.0
numpy==2.2.2
Pillow==11.1.0
scikit-image==0.25.0
tqdm==4.67.1
